src/test/java/com/taskmanager/myapp/controller/DepartmentsControllerTest.java
package com.taskmanager.myapp.controller;

import com.taskmanager.myapp.service.DepartmentsService;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.web.servlet.MockMvc;

import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ActiveProfiles("test")
@ExtendWith(MockitoExtension.class)
@WebMvcTest(DepartmentsController.class)
class DepartmentsControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private DepartmentsService departmentsService;

    @InjectMocks
    private DepartmentsController departmentsController;

    @Test
    @DisplayName("부서 목록 조회")
    void 부서_목록_조회() throws Exception {
        mockMvc.perform(get("/api/departments")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andDo(print());

        verify(departmentsService, times(1)).getAllDepartments();

    }
}